+++ docs/src/cljs/om_bootstrap/docs/components.cljs
@@ -7,6 +7,7 @@
             [om-bootstrap.grid :as g]
             [om-bootstrap.input :as i]
             [om-bootstrap.mixins :as m]
+            [om-bootstrap.modal :as md]
             [om-bootstrap.nav :as n]
             [om-bootstrap.pagination :as pg]
             [om-bootstrap.panel :as p]
@@ -215,6 +216,11 @@
    particular context by adding a " (d/code ":bs-style") " prop.")
    (->example (slurp-example "panel/contextual"))
 
+   (d/h3 "Collapsible panels")
+   (d/p "This panel is collapsed by default and can be extended by clicking
+   on the title")
+   (->example (slurp-example "panel/collapsible"))
+
    (d/h3 "Controlled PanelGroups")
    (d/p (d/code "p/panel-group") "s can be controlled by a parent
    component. The " (d/code ":active-key") " prop dictates which panel
@@ -236,17 +242,11 @@
    (d/p "A rendered modal with header, body, and set of actions in the footer.")
    (d/p "The header is added automatically if you pass in
    a " (d/code ":title") " option.")
-   (TODO)
+   (->example (slurp-example "modal/static"))
 
    (d/h3 "Live Demo")
-   (d/p "Use " (d/code "overlay-trigger") " to create a real modal
-   that's added to the document body when opened.")
-   (TODO)
-
-   (d/h3 "Custom trigger")
-   (d/p "Use " (d/code "overlay-mixin") " in a custom component to
-   manage the modal's state yourself.")
-   (TODO)))
+   (d/p "Build your own component to trigger a modal")
+   (->example (slurp-example "modal/live"))))
 
 ;; ## Tooltips
 
@@ -315,6 +315,7 @@
 
 ;; ## Progress Bars
 
+
 (defn progress-bar-block []
   (section
    "progress"
@@ -324,37 +325,37 @@
         action with simple yet flexible progress bars.")
    (d/h3 "Basic example")
    (d/p "Default progress bar.")
-   (TODO)
+   (->example (slurp-example "progressbar/basic"))
 
    (d/h3 "With label")
    (d/p "Add a " (d/code ":label") " prop to show a visible
    percentage. For low percentages, consider adding
    a" (d/code ":min-width") " to ensure the label's text is fully
    visible.")
-   (TODO)
+   (->example (slurp-example "progressbar/label"))
 
    (d/h3 "Screenreader only label")
    (d/p "Add the " (d/code ":sr-only? true") " option to hide the
    label visually.")
-   (TODO)
+   (->example (slurp-example "progressbar/sr_only_label"))
 
    (d/h3 "Contextual alternatives")
    (d/p "Progress bars use some of the same button and alert classes
    for consistent styles.")
-   (TODO)
+   (->example (slurp-example "progressbar/contextual"))
 
    (d/h3 "Striped")
    (d/p "Uses a gradient to create a striped effect. Not available in IE8.")
-   (TODO)
+   (->example (slurp-example "progressbar/striped"))
 
    (d/h3 "Animated")
    (d/p "Add the " (d/code ":active? true") " option to animate the
    stripes right to left. Not available in IE9 and below.")
-   (TODO)
+   (->example (slurp-example "progressbar/active"))
 
    (d/h3 "Stacked")
    (d/p "Nest " (d/code "pb/progress-bar") "s to stack them.")
-   (TODO)))
+   (->example (slurp-example "progressbar/stacked"))))
 
 ;; ## Navs