+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.5 (2/2/2015)
+
+- Modal component, thanks to @dignati! (https://github.com/racehub/om-bootstrap/pull/50)
+- Support for collapsible panels, thanks to @dignati (https://github.com/racehub/om-bootstrap/pull/45)
+
 ## 0.3.4 (1/27/2015)
 
 - Turned Om into a provided dependency (https://github.com/racehub/om-bootstrap/pull/44)